% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Graph_Input.R
\name{censor_family_onsets_per_family}
\alias{censor_family_onsets_per_family}
\title{wrapper around censor_family_onsets
This functions accepts a tibble with family graphs and each individual. It then censors the onset times for each individual based on the proband's end of follow-up.
Returns a formatted output.}
\usage{
censor_family_onsets_per_family(
  family_graphs,
  tbl,
  start,
  end,
  event,
  status_col,
  aod_col,
  age_eof_col,
  fam_graph_col = "fam_graph",
  fid = "fid",
  pid = "pid",
  merge_by = pid
)
}
\arguments{
\item{family_graphs}{tibble with fid and family graphs columns}

\item{tbl}{tibble with information on each considered individual}

\item{start}{column name of start of follow up, typically date of birth}

\item{end}{column name of the personalised end of follow up}

\item{event}{column name of the event}

\item{status_col}{column name of the status (to be created)}

\item{aod_col}{column name of the age of diagnosis}

\item{age_eof_col}{column name of the age at the end of follow up}

\item{fam_graph_col}{column name of family graphs in the 'family_graphs' object}

\item{fid}{family id, typically the name of the proband that a family graph is centred on}

\item{pid}{personal identifier for each individual}

\item{merge_by}{column names to merge by. If different names are used for family graphs and tbl, a named vector can be specified: setNames(c("id"), c("pid")). Note id is the column name in tbl and pid is the column name in family_graphs.}
}
\value{
A tibble with family ids and updated status, age of diagnosis, and age at end of follow-up for each individual in the family based on the proband's end of follow-up.
}
\description{
wrapper around censor_family_onsets
This functions accepts a tibble with family graphs and each individual. It then censors the onset times for each individual based on the proband's end of follow-up.
Returns a formatted output.
}
\examples{
see vignettes
}
