% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Functions.R
\name{extract_estimation_info_graph}
\alias{extract_estimation_info_graph}
\title{Title Internal Function used to extact input needed from graph input for liability estimation}
\usage{
extract_estimation_info_graph(cur_fam_graph, cur_fid, h2, pid, add_ind = TRUE)
}
\arguments{
\item{cur_fam_graph}{neightbourhood graph of degree n around proband}

\item{cur_fid}{proband ID}

\item{h2}{heritability value from estimate_liability}

\item{pid}{Name of column of personal ID}

\item{add_ind}{Whether the genetic liability be added. Default is TRUE.}
}
\value{
list with two elements: tbl (tibble with all relevant information) and cov (covariance matrix) estimated through graph_based_covariance_construction()
}
\description{
Title Internal Function used to extact input needed from graph input for liability estimation
}
