% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Functions.R
\name{extract_estimation_info_graph_multi}
\alias{extract_estimation_info_graph_multi}
\title{Title Internal Function used to extact input needed from multi-trait graph input for liability estimation}
\usage{
extract_estimation_info_graph_multi(
  cur_fam_graph,
  fid,
  pid,
  cur_fid,
  h2_vec,
  genetic_corrmat,
  phen_names,
  useMixture,
  add_ind = TRUE
)
}
\arguments{
\item{cur_fam_graph}{neightbourhood graph of degree n around proband}

\item{fid}{Name of column of family ID}

\item{pid}{Name of column of personal ID}

\item{cur_fid}{proband ID}

\item{h2_vec}{vector of heritability values from estimate_liability}

\item{genetic_corrmat}{genetic correlation matrix as given to estimate_liability}

\item{phen_names}{vector of phenotype names as given to estimate_liability}

\item{useMixture}{whether mixture model is used}

\item{add_ind}{Whether the genetic liability be added. Default is TRUE.}
}
\value{
list with three elements: tbl (tibble with all relevant information),
cov (covariance matrix) estimated through graph_based_covariance_construction_multi(),
and newOrder (order of individuals in covariance matrix)
}
\description{
Title Internal Function used to extact input needed from multi-trait graph input for liability estimation
}
