% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Functions.R
\name{extract_estimation_info_tbl}
\alias{extract_estimation_info_tbl}
\title{Title Internal Function used to extact input needed for liability estimation}
\usage{
extract_estimation_info_tbl(.tbl, cur_fid, h2, fid, pid, role, add_ind = TRUE)
}
\arguments{
\item{.tbl}{.tbl input from estimate_liability}

\item{cur_fid}{current family ID being worked on}

\item{h2}{heritability value from estimate_liability}

\item{fid}{name of family ID column}

\item{pid}{name of personal ID column}

\item{role}{name of role column}

\item{add_ind}{Whether the genetic liability be added. Default is TRUE.}
}
\value{
list with two elements: tbl (tibble with all relevant information) and cov (covariance matrix) estimated through construct_covmat()
}
\description{
Title Internal Function used to extact input needed for liability estimation
}
