% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Graph_Input.R
\name{familywise_censoring_multi}
\alias{familywise_censoring_multi}
\title{Familywise Censoring for Multiple Outcomes}
\usage{
familywise_censoring_multi(
  family_graphs,
  tbl,
  start,
  end_base,
  phen_names,
  target_outcome = NULL,
  status_col_base = "status",
  aod_col_base = "aod",
  age_eof_col_base = "age_eof",
  fam_graph_col = "fam_graph",
  fid = "fid",
  pid = "pid",
  simplify = TRUE,
  merge_by = pid
)
}
\arguments{
\item{family_graphs}{A tibble containing family graphs for each family.}

\item{tbl}{A tibble containing individual-level data with multiple outcomes.}

\item{start}{The column name representing the start of follow-up time.}

\item{end_base}{The base name for the end of follow-up time columns (e.g., "end" for columns like "end_outcome1", "end_outcome2").}

\item{phen_names}{A vector of phenotype names corresponding to the different outcomes.}

\item{target_outcome}{An optional string specifying the target outcome for censoring. If provided, all outcomes will be censored based on this outcome's end of follow-up time.}

\item{status_col_base}{The base name for the status columns (default is "status").}

\item{aod_col_base}{The base name for the age of diagnosis columns (default is "aod").}

\item{age_eof_col_base}{The base name for the age at end of follow-up columns (default is "age_eof").}

\item{fam_graph_col}{The column name in `family_graphs` that contains the family graph (default is "fam_graph").}

\item{fid}{The column name representing family ID (default is "fid").}

\item{pid}{The column name representing individual ID (default is "pid").}

\item{simplify}{A logical indicating whether to simplify the output by removing columns not specific to an outcome (default is TRUE).}

\item{merge_by}{The column name(s) to merge results by (default is `pid`).}
}
\value{
A tibble containing the familywise censored results for all outcomes.
}
\description{
This function applies familywise censoring across multiple outcomes in a dataset.
If a target outcome is specified, all outcomes are censored based on the end of follow-up time of the target outcome,
then familywise censoring is performed for each outcome individually.
}
\examples{
# TODO: Add examples

}
