% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kendler_functions.R
\name{kendler_simplified}
\alias{kendler_simplified}
\title{(Simplified) Kendler's FGRS}
\usage{
kendler_simplified(
  .tbl = NULL,
  family_graphs = NULL,
  family_graphs_col = "fam_graph",
  pid = "pid",
  fid = "fid",
  role = NULL,
  dadcol,
  momcol,
  env_cor_sib = 0,
  env_cor_f = 0,
  env_cor_m = 0
)
}
\arguments{
\item{.tbl}{A matrix, list or data frame that can be converted into a tibble.
Must have at least five columns that hold the family identifier, the personal
identifier, the role and the lower and upper thresholds. Note that the
role must be one of the following abbreviations
\itemize{
 \item \code{g} (Genetic component of full liability)
 \item \code{o} (Full liability)
 \item \code{m} (Mother)
 \item \code{f} (Father)
 \item \code{c[0-9]*.[0-9]*} (Children)
 \item \code{mgm} (Maternal grandmother)
 \item \code{mgf} (Maternal grandfather)
 \item \code{pgm} (Paternal grandmother)
 \item \code{pgf} (Paternal grandfather)
 \item \code{s[0-9]*} (Full siblings)
 \item \code{mhs[0-9]*} (Half-siblings - maternal side)
 \item \code{phs[0-9]*} (Half-siblings - paternal side)
 \item \code{mau[0-9]*} (Aunts/Uncles - maternal side)
 \item \code{pau[0-9]*} (Aunts/Uncles - paternal side).
}
Defaults to \code{NULL}. If \code{.tbl} is provided, \code{family_graphs} must be \code{NULL}.}

\item{family_graphs}{A tibble with columns pid and family_graph_col, dadcol, and momcol.
See prepare_graph for construction of the graphs. The family graphs Defaults to NULL.}

\item{family_graphs_col}{Name of column with family graphs in family_graphs. Defaults to "fam_graph".}

\item{pid}{A string holding the name of the column in \code{.tbl} (or \code{family} and
\code{threshs}) that hold the personal identifier(s). Defaults to "PID".}

\item{fid}{A string holding the name of the column in \code{.tbl} or \code{family} that
holds the family identifier. Defaults to "fid".}

\item{role}{A string holding the name of the column in \code{.tbl} that
holds the role. Each role must be chosen from the following list of abbreviations
\itemize{
 \item \code{g} (Genetic component of full liability)
 \item \code{o} (Full liability)
 \item \code{m} (Mother)
 \item \code{f} (Father)
 \item \code{c[0-9]*.[0-9]*} (Children)
 \item \code{mgm} (Maternal grandmother)
 \item \code{mgf} (Maternal grandfather)
 \item \code{pgm} (Paternal grandmother)
 \item \code{pgf} (Paternal grandfather)
 \item \code{s[0-9]*} (Full siblings)
 \item \code{mhs[0-9]*} (Half-siblings - maternal side)
 \item \code{phs[0-9]*} (Half-siblings - paternal side)
 \item \code{mau[0-9]*} (Aunts/Uncles - maternal side)
 \item \code{pau[0-9]*} (Aunts/Uncles - paternal side).
}
Defaults to "role".}

\item{dadcol}{column name of father in family_graphs or .tbl.}

\item{momcol}{column name of mother in family_graphs or .tbl.}

\item{env_cor_sib}{Cohabitation effect, i.e. Factor by which the siblings are weighted. Defaults to 1.}

\item{env_cor_f}{Cohabitation effect, i.e. Factor by which the father is weighted. Defaults to 1.}

\item{env_cor_m}{Cohabitation effect, i.e. Factor by which the mother is weighted. Defaults to 1.}
}
\value{
A tibble with summary values used to calculate the simplified kendler FGRS and the FGRS itself.
}
\description{
Function to calculate the simplified version of Kendler's FGRS based on family data.
}
\details{
The coding of the cohabitation effects differ slightly from the one suggested by Kendler et al.
Here, it is coded as env_cor_\* = env_eff_\* / (gen_eff_\* + env_eff_\*), while the original implementation
suggestes coding it as gen_eff_\*  / (gen_eff_\* + env_eff_\*), i.e. the two are related as 1 - env_cor_\*.
}
\examples{
# See Vignettes.
}
