% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelleing_and_plotting_functions.R
\name{label_relatives}
\alias{label_relatives}
\title{Label Pairwise Relationships Based on Generational Distance and Kinship Coefficient}
\usage{
label_relatives(tbl)
}
\arguments{
\item{tbl}{A tibble or data frame containing at least the following columns:
\describe{
  \item{fid}{Column with family identifier (typically the proband's id).}
  \item{id1}{Identifier for the first individual.}
  \item{id2}{Identifier for the second individual.}
  \item{gen.x}{Number of generations between `id1` and their most recent common ancestor with `id2`.}
  \item{gen.y}{Number of generations between `id2` and their most recent common ancestor with `id1`.}
  \item{k}{Estimated kinship coefficient between the two individuals.}
}}
}
\value{
A tibble with the following columns:
\describe{
  \item{fid}{Column with family identifier (typically the proband's id).}
  \item{id1}{Identifier for the first individual.}
  \item{id2}{Identifier for the second individual.}
  \item{gen.x}{Generational distance for \code{id1}.}
  \item{gen.y}{Generational distance for \code{id2}.}
  \item{k}{Kinship coefficient between the two individuals.}
  \item{lab}{Assigned relationship label (e.g., `"S"`, `"P"`, `"1C"`, `"H1C"`, `"2GP"`, etc.).}
}
}
\description{
Assigns standard pedigree relationship labels (e.g., *Parent*, *Child*, *Sibling*, *Grandparent*, *Cousin*)
to all pairs of individuals based on their generational distances (`gen.x`, `gen.y`)
and kinship coefficients (`k`), typically produced by [get_generations()].
}
\details{
This function derives descriptive relationship labels using generational differences
and kinship patterns. The labels are written in a short-hand notation, an explaination of a subset is given below:
\itemize{
  \item *P* - Parent
  \item *Ch* - Child
  \item *S* - Sibling
  \item *GP* - Grandparent
  \item *Pib* - "Pibling" (parental sibling; aunt/uncle)
  \item *Nib* - "Nibling" (sibling's child; niece/nephew)
  \item *GCh* - Grandchild
  \item *GPib* - Grandpibling (grandparent's sibling)
  \item *GNib* - Grandnibling (sibling's grandchild)
  \item *C* - Cousin
  \item *1C1R* - First Cousin Once Removed
  \item *2C2R* - Second Cousin Twice Removed
  \item *H* prefix - Half relationships (e.g., *HS* for Half-Sibling)
}
}
\examples{
# see vignette on identifying and labelling relatives

}
\seealso{
[get_generations()] for computing the generational and kinship inputs used by this function.
}
