% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFGRS-package.R
\docType{package}
\name{LTFGRS-package}
\alias{LTFGRS}
\alias{LTFGRS-package}
\title{LTFGRS: Implementation of Several Phenotype-Based Family Genetic Risk Scores}
\description{
Implementation of several phenotype-based family genetic risk scores with unified input data and data preparation functions to help facilitate the required data preparation and management. The implemented family genetic risk scores are the extended liability threshold model conditional on family history (LT-FH++) from Pedersen (2022) \doi{10.1016/j.ajhg.2022.01.009} and Pedersen (2023) \url{https://www.nature.com/articles/s41467-023-41210-z}, Pearson-Aitken Family Genetic Risk Scores (PA-FGRS) from Krebs (2024) \doi{10.1016/j.ajhg.2024.09.009}, and family genetic risk score by Kendler (2021) \doi{10.1001/jamapsychiatry.2021.0336}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emilmip.github.io/LTFGRS/}
  \item Report bugs at \url{https://github.com/EmilMiP/LTFGRS/issues}
}

}
\author{
\strong{Maintainer}: Emil Michael Pedersen \email{emp@ph.au.dk}

Authors:
\itemize{
  \item Jette Steinbach \email{jst@econ.au.dk}
  \item Morten Dybdahl Krebs \email{morten.dybdahl.krebs@regionh.dk}
}

Other contributors:
\itemize{
  \item Lucas Rasmussen \email{lar.ncrr@au.dk} [contributor]
}

}
\keyword{internal}
