% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Graph_Input.R
\name{prepare_thresholds_multi}
\alias{prepare_thresholds_multi}
\title{Prepare thresholds for multiple phenotypes}
\usage{
prepare_thresholds_multi(
  .tbl,
  CIP_list,
  phen_names,
  CIP_merge_columns = c("sex", "birth_year", "age"),
  CIP_cip_col = "cip",
  age_col = "age",
  age_eof_base = "age_eof",
  status_col_base = "status",
  lower_base = "lower",
  upper_base = "upper",
  thr_col = "thr",
  K_i_col = "K_i",
  K_pop_col = "K_pop",
  Kpop_list = "useMax",
  personal_thr = TRUE,
  lower_equal_upper = FALSE,
  interpolation = "xgboost",
  bst.params = list(max_depth = 10, base_score = 0, nthread = 4, min_child_weight = 10),
  min_CIP_value = 1e-05,
  xgboost_itr = 30
)
}
\arguments{
\item{.tbl}{Tibble with family and personal id columns, as well as CIP_merge_columns and status for each phenotype.}

\item{CIP_list}{List of tibbles with population representative cumulative incidence proportions. Each tibble must contain columns from \code{CIP_merge_columns} and \code{cIP_cip_col}.}

\item{phen_names}{Vector of phenotype names. Used to identify status columns and to name output columns.}

\item{CIP_merge_columns}{The columns the CIPs are subset by, e.g. CIPs by birth_year, sex. and age_col.}

\item{CIP_cip_col}{Name of column with CIP values.}

\item{age_col}{Name of column with age at the end of follow-up or age at diagnosis for cases.}

\item{age_eof_base}{Base name of age at end of follow-up column. The actual column name is constructed by appending the phenotype name. Defaults to "age_eof".}

\item{status_col_base}{Base name of status column. The actual column name is constructed by appending the phenotype name. Defaults to "status".}

\item{lower_base}{Base name of lower threshold column. The actual column name is constructed by appending the phenotype name. Defaults to "lower".}

\item{upper_base}{Base name of upper threshold column. The actual column name is constructed by appending the phenotype name. Defaults to "upper".}

\item{thr_col}{Base name of threshold column. The actual column name is constructed by appending the phenotype name. Defaults to "thr".}

\item{K_i_col}{Base name of individual cumulative incidence proportion column. The actual column name is constructed by appending the phenotype name. Defaults to "K_i".}

\item{K_pop_col}{Base name of population prevalence column. The actual column name is constructed by appending the phenotype name. Defaults to "K_pop".}

\item{Kpop_list}{List of population prevalence tibbles or "useMax" for each phenotype. If a tibble is provided, it must contain columns from \code{.tbl} and a column named "K_pop" with population prevalence values. Defaults to "UseMax".}

\item{personal_thr}{Should thresholds be based on stratified CIPs or population prevalence?}

\item{lower_equal_upper}{Should the upper and lower threshold be the same for cases? Can be used if CIPs are detailed, e.g. stratified by birth year and sex.}

\item{interpolation}{Type of interpolation, defaults to "xgboost".}

\item{bst.params}{List of parameters to pass on to xgboost. See xgboost documentation for details.}

\item{min_CIP_value}{Minimum cip value to allow. Too low values may lead to numerical instabilities.}

\item{xgboost_itr}{Number of iterations to run xgboost for.}
}
\value{
Tibble with (personlised) thresholds for each family member (lower & upper), the calculated cumulative incidence proportion for each individual (K_i), and population prevalence within an individuals CIP strata (K_pop; max value in stratum) for each phenotype. The threshold and other potentially relevant information can be added to the family graphs with \code{familywise_attach_attributes}.
}
\description{
This function is a wrapper around \code{prepare_thresholds} to prepare thresholds for multiple phenotypes at once.
}
\examples{
# TODO: create simple example

}
