% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{tnorm_var}
\alias{tnorm_var}
\title{Title: Calculate the variance of the truncated normal distribution}
\usage{
tnorm_var(mu = 0, sigma = 1, lower = -Inf, upper = Inf)
}
\arguments{
\item{mu}{mean value of normal distribution}

\item{sigma}{standard deviation of normal distribution}

\item{lower}{lower threshold}

\item{upper}{upper threshold}
}
\value{
mean value of the truncated normal distribution
}
\description{
Title: Calculate the variance of the truncated normal distribution
}
\examples{
tnorm_var()
}
