% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Graph_Input.R
\name{assign_family_specific_thresholds}
\alias{assign_family_specific_thresholds}
\title{Title}
\usage{
assign_family_specific_thresholds(
  cur_fam_graph,
  cur_proband,
  fam_id_col,
  attr_tbl,
  attr_names,
  censor_proband_thrs = TRUE
)
}
\arguments{
\item{cur_fam_graph}{igraph object (neighbourhood graph around a proband) with family members of degree n}

\item{cur_proband}{current proband id (center of the neighbourhood graph)}

\item{fam_id_col}{column name of family id}

\item{attr_tbl}{tibble with family id and attributes for each family member}

\item{attr_names}{names of attributes to be assigned to each node (family member) in the graph}

\item{censor_proband_thrs}{should proband thresholds be censored? Defaults to TRUE. Used proband's information for prediction.}
}
\value{
igraph object (neighbourhood graph around a proband) with updated attributes for each node in the graph
}
\description{
Title
}
