% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Graph_Input.R
\name{get_family_graphs}
\alias{get_family_graphs}
\title{Identify family members of degree n}
\usage{
get_family_graphs(
  pop_graph,
  ndegree,
  proband_vec,
  fid = "fid",
  fam_graph_col = "fam_graph",
  mindist = 0,
  mode = "all"
)
}
\arguments{
\item{pop_graph}{population graph from prepare_graph()}

\item{ndegree}{number of steps away from proband to include}

\item{proband_vec}{vector of proband ids to create family graphs for}

\item{fid}{column name of proband ids in the output}

\item{fam_graph_col}{column name of family graphs in the output}

\item{mindist}{minimum distance from proband to include in the graph (experimental, untested), defaults to 0, passed directly to make_neighborhood_graph}

\item{mode}{type of distance measure in the graph (experimental, untested), defaults to "all", passed directly to make_neighborhood_graph}
}
\value{
tibble with two columns, family ids and family graphs
}
\description{
This function identifies individuals ndegree-steps away from the proband in the population graph
}
\examples{
See vignettes
}
