% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFGRS-package.R
\docType{package}
\name{LTFGRS-package}
\alias{LTFGRS}
\alias{LTFGRS-package}
\title{LTFGRS: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emilmip.github.io/LTFGRS/}
  \item Report bugs at \url{https://github.com/EmilMiP/LTFGRS/issues}
}

}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
