% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTFGRS-package.R
\docType{package}
\name{LTFGRS-package}
\alias{LTFGRS}
\alias{LTFGRS-package}
\title{LTFGRS: Implementation of several phenotype-based family history genetic predictors}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emilmip.github.io/LTFGRS/}
  \item Report bugs at \url{https://github.com/EmilMiP/LTFGRS/issues}
}

}
\author{
\strong{Maintainer}: Emil Michael Pedersen \email{emp@ph.au.dk}

Authors:
\itemize{
  \item Jette Steinbach \email{jst@econ.au.dk}
}

Other contributors:
\itemize{
  \item Lucas Rasmussen \email{lar.ncrr@au.dk} [contributor]
}

}
\keyword{internal}
